requirements.txt
requests>=2.31.0
pandas>=2.0.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
streamlit>=1.28.0
plotly>=5.18.0
ollama>=0.1.0
pillow>=10.0.0
